client/requirements.txt
websocket-client
rel
pyautogui
colorama
pyinstaller